readme: add link to GitLab ME #4892
I do not understand how protected branches and closing merge requests
from the command line interact. So far I could see MR !8 being closed
automatically just seconds after having pushed the exact same commit to
master. But master was not a protected branch.

It is not detected when a rebase + squash of an outdated merge request
would result in the same tree as push to master.