ci: reject merge requests that are not clang-formatted
Assume the destination branch to always be master because there are no
suitable environment variables. ${CI_MERGE_REQUEST_TARGET_BRANCH_NAME}
is not set and thus not usable unless the 'only: merge_requests' YAML
element is set. This would have to be set on all jobs, otherwise the
clang-format job would be the only one executed for merge requests.

Colorize the difference making sure to be compatible to dash, i.e.
subscripts of the form `${line:0:2}` are not available.
An alternative would be colordiff which is not currently available in
the builder image.