clang-tidy: treat most warnings as errors
Disable llvm-header-guard because it is too specific to LLVM and not
configurable. Do not treat warnings as errors that cannot be influenced.
The NOLINT and NOLINTNEXTLINE comments are not suitable because warnings
related to Boost.Test should be disabled on a per-file level.
Do not filter any headers from the output.

Reflect a change in GitLab CI between 11.8.0 and 11.9.0-rc2 where remote
branches no longer get their local counterparts